python/requirements.txt
python-dotenv==0.19.2
redis==5.0.0